fix sidebar and dashboard problem

+++ amana_frontend/src/pages/Dashboard.jsx
@@ -63,7 +63,7 @@ export default function Dashboard() {
     const lineData = stats?.par_mois?.map((m) => ({
         month: m.mois,
         crbt: parseFloat(m.total_crbt),
-        envois: parseFloat(m.total_envois)
+        envois: parseInt(m.total_envois)
     })) || [];
 
     const payePercent = stats
